+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.1.1 (2015-04-28)
+
+
+#### Features
+
+* **filter:** provide filter for use on interpolation ([007a1e75](https://github.com/awerlang/angular-easy-masks/commit/007a1e75))
+* **options:** create option to remove separators ([1e6890d5](https://github.com/awerlang/angular-easy-masks/commit/1e6890d5))
+* **registry:** create registry for commonly used masks ([9256b916](https://github.com/awerlang/angular-easy-masks/commit/9256b916))
+
+
 ## 0.1.0 (2015-04-27)